fix(booking-service): address all review comments for S3-F6 analytics feature (55-24701)
- Update BookingAnalyticsDTO to use BigDecimal for monetary fields
- Fix BigDecimal type conversion in service method for monetary values
- Add date validation for analytics endpoint (startDate <= endDate)
- Use LocalTime.MAX for end-of-day calculation to include full day
- Add proper imports instead of fully qualified names
- Improve method signatures and return types

+++ booking-service/src/main/java/com/team28/booking/booking/dto/BookingAnalyticsDTO.java
@@ -1,11 +1,13 @@
 package com.team28.booking.booking.dto;
 
+import java.math.BigDecimal;
+
 public record BookingAnalyticsDTO(
         long totalBookings,
         long completedBookings,
         long cancelledBookings,
-        double totalRevenue,
-        double averageBookingPrice,
+        BigDecimal totalRevenue,
+        BigDecimal averageBookingPrice,
         double completionRate
 ) {
 }
